+++ NEWS.md
@@ -2,4 +2,4 @@
 
 - Moved to C++ since I hate string handling in C
 - Made the lookup routing code more consistent (function calls for each type)
-- Added support for TXT & CNAME records
+- Added support for TXT, SOA & CNAME records